requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langgraph
chromadb
pypdf
sentence-transformers
uvicorn
gradio
python-dotenv
